download sqlserver 2008 enterprise r2 install package.

+++ module.ps1
@@ -1,9 +1,12 @@
 $root = Split-Path -Parent $MyInvocation.MyCommand.Path
 $moduleDir = "$root\ChefDotNet"
 
-rm $moduleDir -Recurse -Force
-mkdir "$moduleDir"
-mkdir "$moduleDir\lib"
+if (Test-Path $moduleDir) { 
+	rm $moduleDir -Recurse -Force 
+	mkdir "$moduleDir"
+	mkdir "$moduleDir\lib"
+}
+
 cp -r "$root\lib\*" "$moduleDir\lib"
 cp *.psd1 $moduleDir
 cp *.psm1 $moduleDir
